refactor(security): split rate-limiter god-object behind facade (S4 of #141)
Pure code movement — no behavior changes. The single `RateLimiter`
(1178 LOC) becomes a thin facade (213 LOC) that forwards every
public static call to one of three focused classes.

  - RateLimitChecker (1073 LOC) — get_settings, check_all, check_*
    (ip/email/cpf/global/device/verification/user_limit), record_*,
    plus all internal helpers (counters, blacklist, whitelist,
    temporary blocks, format/window utilities, get_user_ip).
  - RateLimitLogger (99 LOC) — log_attempt, cleanup_old_logs (private),
    cleanup_expired (cron entry point).
  - RateLimitStats (44 LOC) — get_stats aggregation.

Cross-class adjustments (no behavior change):
  - 9 `self::log_attempt(...)` call sites in `check_all` and
    `check_verification` rewritten to `RateLimitLogger::log_attempt(...)`.
  - `get_user_ip()` promoted from private to public so
    `RateLimitLogger::log_attempt` can call it as
    `RateLimitChecker::get_user_ip()` (alternative would have been to
    duplicate the helper, violating the no-logic-changes rule).
  - `CACHE_GROUP` constant kept on the facade — many external callers
    reference `RateLimiter::CACHE_GROUP` directly. Methods inside the
    new classes use `RateLimiter::CACHE_GROUP` (a deliberate one-way
    dependency that preserves the public API).

Test harness updated (4 files): the reflection-based `$settings_cache`
reset in setUp() now targets `RateLimitChecker` (the property's new
home) instead of `RateLimiter`. Two of the test files used
`hasProperty` guards that would have silently no-op'd after the move
and leaked settings between tests — repointing the guard restores
their original protective behavior.

Public API of `RateLimiter` is byte-identical: every external caller
(13 production files + tests) keeps working without any change.

Deviation: `RateLimitChecker` is 1073 LOC, above the issue's <800-LOC
target. The three-class split named in the issue spec (Checker /
Logger / Stats) puts every `check_*` method in Checker by design;
hitting <800 would require further sub-splitting (e.g. extracting
Device-fingerprint or Blocklist sub-classes), which deviates from the
explicit issue plan more than the LOC overrun does. Flagging here so
this can be revisited if the LOC bar is the binding constraint.

Verification:
  - php -l: clean on all 4 files.
  - vendor/bin/phpunit: 3870/3870 OK.
  - vendor/bin/phpstan analyse includes/security/: no errors.
