+++ tests/Unit/PublicCsvDownloadTest.php
@@ -11,6 +11,7 @@ use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use PHPUnit\Framework\TestCase;
 use FreeFormCertificate\Frontend\PublicCsvDownload;
 use FreeFormCertificate\Security\RateLimiter;
+use FreeFormCertificate\Security\RateLimitChecker;
 
 /**
  * Tests for PublicCsvDownload: shortcode + admin-post handler.
@@ -46,8 +47,9 @@ class PublicCsvDownloadTest extends TestCase {
         $this->meta_store   = array();
         $this->meta_updates = array();
 
-        // Reset RateLimiter's static settings cache between tests.
-        $rl = new \ReflectionClass( RateLimiter::class );
+        // Reset RateLimiter's static settings cache between tests. The cache
+        // moved to RateLimitChecker in the S4 facade refactor.
+        $rl = new \ReflectionClass( RateLimitChecker::class );
         if ( $rl->hasProperty( 'settings_cache' ) ) {
             $prop = $rl->getProperty( 'settings_cache' );
             $prop->setAccessible( true );
